# Add context authorized rest api feature