--- a/includes/API/GetPageByRevIdHandler.php
+++ b/includes/API/GetPageByRevIdHandler.php
@@ -45,7 +45,7 @@ class GetPageByRevIdHandler extends ContextAuthorized {
 		);
 
 		if ( !$res ) {
-			throw new HttpException( "rev_id not found in the database", 404 );
+			throw new HttpException( "Not found", 404 );
 		}
 
 		return [