--- a/includes/API/GetServerInfoHandler.php
+++ b/includes/API/GetServerInfoHandler.php
@@ -14,4 +14,14 @@ class GetServerInfoHandler extends SimpleHandler {
 		$apiVersion = ServerInfo::DA_API_VERSION;
 		return [ 'api_version' => $apiVersion ];
 	}
+
+	/** @inheritDoc */
+	public function needsReadAccess() {
+		return false;
+	}
+
+	/** @inheritDoc */
+	public function needsWriteAccess() {
+		return false;
+	}
 }