--- a/includes/API/GetWitnessDataHandler.php
+++ b/includes/API/GetWitnessDataHandler.php
@@ -27,6 +27,11 @@ class GetWitnessDataHandler extends SimpleHandler {
 	}
 
 	/** @inheritDoc */
+	public function needsReadAccess() {
+		return false;
+	}
+
+	/** @inheritDoc */
 	public function getParamSettings() {
 		return [
 			'witness_event_id' => [