--- a/includes/API/GetWitnessDataHandler.php
+++ b/includes/API/GetWitnessDataHandler.php
@@ -16,7 +16,7 @@ class GetWitnessDataHandler extends SimpleHandler {
 		#Expects 'get_witness_data\'- USES witness_event_id - used to retrieve all required data to execute a witness event (including domain_manifest_verification_hash, merkle_root, network ID or name, witness smart contract address, transaction_id) for the publishing via Metamask'];
 		$output = getWitnessData( $witness_event_id );
 		if ( empty( $output ) ) {
-			throw new HttpException( "witness_event_id not found in the witness_events table of the database.", 404 );
+			throw new HttpException( "Not found", 404 );
 		}
 		return $output;
 	}