--- a/includes/API/RequestHashHandler.php
+++ b/includes/API/RequestHashHandler.php
@@ -66,7 +66,7 @@ class RequestHashHandler extends ContextAuthorized {
 	protected function provideTitle( int $revId ): ?Title {
 		$revisionRecord = $this->revisionLookup->getRevisionById( $revId );
 		if ( !$revisionRecord ) {
-			throw new HttpException( "invalid rev_id", 404 );
+			throw new HttpException( "Not found", 404 );
 		}
 		return $revisionRecord->getPageAsLinkTarget();
 	}