--- a/includes/API/VerifyPageHandler.php
+++ b/includes/API/VerifyPageHandler.php
@@ -90,7 +90,7 @@ class VerifyPageHandler extends ContextAuthorized {
 	protected function provideTitle( int $revId ): ?Title {
 		$revisionRecord = $this->revisionLookup->getRevisionById( $revId );
 		if ( !$revisionRecord ) {
-			throw new HttpException( "invalid rev_id", 404 );
+			throw new HttpException( "Not found", 404 );
 		}
 		return $revisionRecord->getPageAsLinkTarget();
 	}