https://github.com/inblockio/DataAccounting/issues/187