@rht 2021-12-07
We also need to measure how much slower the external verifier take to verify a page. Maybe 10% slower?