@rht 2021-12-07
Regarding the complexity of the permission checks, am I right to say that they are about as complex as the write API permission checks?