@pwirth 2021-12-07
> 

i would say, that permission checks on single pages are more complex then just the global check form the write API's