@pwirth 2021-12-07
> We also need to measure how much slower the external verifier take to verify a page [with this PR]. Maybe 10% slower?

its at least either 1 (title from name) or 2 (revision from revision id and title from revision) more queries. i dont know how we could messure the actual impact on the performance