@rht 2021-12-07
We can measure by using `time ./verifier.js 'Main Page'` (via https://github.com/inblockio/data-accounting-external-verifier), for a 'Main Page' page that has at least 10 or 20 revisions.