@rht 2021-12-09
This is the benchmark using hyperfine, with 10 revisions.
Current main:
```
Benchmark 1: ./verifier.js 'Main Page'
  Time (mean ± σ):      1.979 s ±  0.026 s    [User: 1.019 s, System: 0.105 s]
  Range (min … max):    1.919 s …  2.028 s    10 runs
```
This revision
```
Benchmark 1: ./verifier.js 'Main Page'
  Time (mean ± σ):      2.037 s ±  0.087 s    [User: 1.012 s, System: 0.108 s]
  Range (min … max):    1.967 s …  2.275 s    10 runs
```
About 3% slower.
I will test with 50 revisions soon.