@rht 2021-12-09
With 50 revisions.
Current main
```
Benchmark 1: ./verifier.js 'Main Page'
  Time (mean ± σ):      6.352 s ±  0.042 s    [User: 1.521 s, System: 0.169 s]
  Range (min … max):    6.279 s …  6.452 s    20 runs
```

This PR
```
Benchmark 1: ./verifier.js 'Main Page'
  Time (mean ± σ):      6.474 s ±  0.030 s    [User: 1.508 s, System: 0.169 s]
  Range (min … max):    6.405 s …  6.525 s    20 runs
```
1.9% slower.