@pwirth 2021-12-09
> ^OK so performance is totally not an issue.

Good to know, thanks for testing!